assistant_core/requirements.txt
fastapi
uvicorn
pydantic
qdrant-client
beautifulsoup4
langchain-text-splitters
google-generativeai
